CacheManager (list+files) works + minor generic change

+++ MatrixProblem.h
@@ -66,10 +66,9 @@ class MatrixProblem : public Searchable<T> {
     for (auto itr = path.begin(); next(itr) != path.end(); itr++) {
       try {
         // get the direction between current two states
-        //string d = getDirection(*itr, *next(itr));
         path_str.append(getDirection(*itr, *next(itr)) + " ");
         // get the current cost
-        path_str.append("(" + to_string(itr->getSumCost()) + ")" + " ");
+        path_str.append("(" + to_string((int) itr->getSumCost()) + ")" + " ");
       } catch (const char *c) {
         cerr << c << endl;
       }
